# go/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  requests.EmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  requests.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  requests.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  requests.SignupRequest:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - address
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /Signup:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/requests.SignupRequest'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Allow you to register as a new User
      tags:
      - Auth
  /forgotten_password:
    post:
      consumes:
      - application/json
      description: Envoie un mail de reset de mot de passe
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Connexion réussie
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Envoie un mail à un user existant afin de réinitialiser son mot de
        passe
      tags:
      - Auth
  /login:
    post:
      consumes:
      - application/json
      description: login to the app
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Connexion réussie
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Allow you to log and have an JWT Token
      tags:
      - Auth
  /logout:
    post:
      consumes:
      - application/json
      description: Delete your token session
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Déconnexion réussie
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Logout
      tags:
      - Auth
  /profile:
    get:
      consumes:
      - application/json
      description: Retourne les informations du profil de l'utilisateur connecté
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Récupère le profil de l'utilisateur actuellement connecté
      tags:
      - Auth
  /reset_password:
    put:
      consumes:
      - application/json
      description: Permet à l'utilisateur de réinitialiser son mot de passe en utilisant
        un jeton de réinitialisation valide.
      parameters:
      - description: Données pour réinitialiser le mot de passe
        in: body
        name: requests.ResetPasswordRequest
        required: true
        schema:
          $ref: '#/definitions/requests.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Mot de passe réinitialisé avec succès
        "400":
          description: Token invalide ou expiré
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/gin.H'
      summary: Réinitialiser le mot de passe
      tags:
      - Auth
  /send_invitation:
    post:
      consumes:
      - application/json
      description: Envoie un mail d'invitation afin de de rejoindre un groupen de
        voyage
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitation envoyée
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - Bearer: []
      summary: Rejoindre un groupe de voyage
      tags:
      - Voyage
swagger: "2.0"